Various bug fixes:
- Fix typo where scopes never had the manage_account_details permission.
- Fix typo where sessions immediately expired.
- Fix bug in can_alter_permissions check.
- Return an error properly when a resource is not found instead of raising a 500 error.
- Explicitly return no response when there is no response.
- Allow creating a user with permissions already set.
- Don't error when credentials are invalid/not present, just give a read-only scope.

+++ polympics_server/routes/utils.py
@@ -48,17 +48,19 @@ def authenticate(
     elif credentials.username.upper().startswith('S'):
         model = Session
     else:
-        return None
+        return Scope()
     try:
         id = int(credentials.username[1:])
     except ValueError:
-        return None
+        return Scope()
     session = model.get_or_none(
         (model.id == id) & (model.token == credentials.password)
     )
+    if not session:
+        return Scope()
     if session.expired:
         session.delete_instance()
-        return None
+        return Scope()
     return session.scope
